+++ src/nvim/ex_cmds2.h
@@ -4,8 +4,8 @@
 #include <stdbool.h>
 
 #include "nvim/ex_docmd.h"
+#include "nvim/runtime.h"
 
-typedef void (*DoInRuntimepathCB)(char_u *, void *);
 
 //
 // flags for check_changed()
@@ -15,10 +15,6 @@ typedef void (*DoInRuntimepathCB)(char_u *, void *);
 #define CCGD_FORCEIT    4       // ! used
 #define CCGD_ALLBUF     8       // may write all buffers
 #define CCGD_EXCMD      16      // may suggest using !
-
-// last argument for do_source()
-#define DOSO_NONE       0
-#define DOSO_VIMRC      1       // loading vimrc file
 
 #ifdef INCLUDE_GENERATED_DECLARATIONS
 # include "ex_cmds2.h.generated.h"
